Fix for code review feedback
