+++ src/cpp/server/utils/platform/path_linux.cpp
@@ -61,7 +61,18 @@ public:
                 }
             }
         }
-        throw std::runtime_error("Unable to resolve writable runtime directory from XDG_RUNTIME_DIR");
+
+        // System services can get a RuntimeDirectory= without XDG_RUNTIME_DIR.
+        if (const char* runtime_dir = std::getenv("RUNTIME_DIRECTORY");
+            runtime_dir && runtime_dir[0] != '\0') {
+            std::error_code ec;
+            fs::path base(runtime_dir);
+            if (fs::is_directory(base, ec) && !ec && access(runtime_dir, W_OK | X_OK) == 0) {
+                return base.string();
+            }
+        }
+
+        throw std::runtime_error("Unable to resolve writable runtime directory from XDG_RUNTIME_DIR or RUNTIME_DIRECTORY");
     }
 
     std::vector<std::string> get_install_prefixes() override {
